Writing binary files for lightcones
