+++ Exec/ParticleFilterTest/DarkMatterParticleContainer.cpp
@@ -7,7 +7,7 @@
 
 using namespace amrex;
 #include <constants_cosmo.H>
-#include <temp.H>
+#include <LightConeParticle.H>
 
 /// These are helper functions used when initializing from a morton-ordered
 /// binary particle file.
@@ -579,6 +579,7 @@ DarkMatterParticleContainer::moveKickDrift (amrex::MultiFab&       acceleration,
             Gpu::streamSynchronize();
         }
     }
+	delete ShellPC;
 }
 
 void
@@ -720,69 +721,69 @@ void store_dm_particle_single (amrex::ParticleContainer<1+AMREX_SPACEDIM, 0>::Su
     amrex::Real dt_a_cur_inv = dt * a_cur_inv;
     const GpuArray<Real,AMREX_SPACEDIM> center({AMREX_D_DECL((phi[0]-plo[0])*0.5,(phi[1]-plo[1])*0.5,(phi[2]-plo[2])*0.5)});
 
-    if (do_move == 1) 
-         {
-            p2.rdata(0)=p.rdata(0);
-            bool result=false;
-            Real lenx = phi[0]-plo[0];
-            Real leny = phi[1]-plo[1];
-            Real lenz = phi[2]-plo[2];
-            int maxind[3]; 
-            maxind[0] = floor((radius_outer+lenx*0.5)/lenx);
-            maxind[1] = floor((radius_outer+leny*0.5)/leny);
-            maxind[2] = floor((radius_outer+lenz*0.5)/lenz);
+    if (do_move == 1) {
+		p2.rdata(0)=p.rdata(0);
+		bool result=false;
+		Real lenx = phi[0]-plo[0];
+		Real leny = phi[1]-plo[1];
+		Real lenz = phi[2]-plo[2];
+		int maxind[3]; 
+		maxind[0] = floor((radius_outer+lenx*0.5)/lenx);
+		maxind[1] = floor((radius_outer+leny*0.5)/leny);
+		maxind[2] = floor((radius_outer+lenz*0.5)/lenz);
 
-            Real xlen, ylen, zlen;
-            //printf("Value is %d\n", maxind);
-	    int local_index=-1;
-        for(int idir=-maxind[0];idir<=maxind[0];idir++)
-            for(int jdir=-maxind[1];jdir<=maxind[1];jdir++)
-                for(int kdir=-maxind[2];kdir<=maxind[2];kdir++)
-                    {
-                        xlen = p.pos(0)+(idir)*(phi[0]-plo[0]) - center[0];
-                        ylen = p.pos(1)+(jdir)*(phi[1]-plo[1]) - center[1];
-                        zlen = p.pos(2)+(kdir)*(phi[2]-plo[2]) - center[2];
-                        Real mag = sqrt(xlen*xlen+ylen*ylen+zlen*zlen);
-                        result=result? true : (mag>radius_inner && mag<radius_outer);
-	      if(int(mag>radius_inner && mag<radius_outer))
-                local_index++;
+		Real xlen, ylen, zlen;
+		int local_index=-1;
+		for(int idir=-maxind[0];idir<=maxind[0];idir++) {
+			for(int jdir=-maxind[1];jdir<=maxind[1];jdir++) {
+				for(int kdir=-maxind[2];kdir<=maxind[2];kdir++) {
+					xlen = p.pos(0)+(idir)*(phi[0]-plo[0]) - center[0];
+					ylen = p.pos(1)+(jdir)*(phi[1]-plo[1]) - center[1];
+					zlen = p.pos(2)+(kdir)*(phi[2]-plo[2]) - center[2];
+					Real mag = sqrt(xlen*xlen+ylen*ylen+zlen*zlen);
+					result=result? true : (mag>radius_inner && mag<radius_outer);
 
-              if(local_index==index) {
-                int comp=0;
-                p2.pos(comp) = p.pos(comp)+(idir)*(phi[comp]-plo[comp]);
-                Real x1 = p2.pos(comp);
-                Real vx = p.rdata(comp+1);
-                comp=1;
-                p2.pos(comp) = p.pos(comp)+(jdir)*(phi[comp]-plo[comp]);
-                Real y1 = p2.pos(comp);
-                Real vy = p.rdata(comp+1);
-                comp=2;
-                p2.pos(comp) = p.pos(comp)+(kdir)*(phi[comp]-plo[comp]);
-                Real z1 = p2.pos(comp);
-                Real vz = p.rdata(comp+1);
+					if(int(mag>radius_inner && mag<radius_outer)){
+						local_index++;
+					}
 
-				// There are two calls to this routine. Do the particle output writing only once 
-				// based on the bool.
-				if(is_file_write) {
-					//fprintf(file_lightcone_csv,"%0.15g, %0.15g, %0.15g, %0.15g, %0.15g, %0.15g \n", x1, y1, z1, vx, vy, vz);
-					myParticle tmp;
-					tmp.x = x1; tmp.y = y1; tmp.z = z1;
-					tmp.vx = vx; tmp.vy = vy; tmp.vz = vz;
-					shell_particles.emplace_back(tmp);
+					if(local_index==index) {
+						int comp=0;
+						p2.pos(comp) = p.pos(comp)+(idir)*(phi[comp]-plo[comp]);
+						Real x1 = p2.pos(comp);
+						Real vx = p.rdata(comp+1);
+						comp=1;
+						p2.pos(comp) = p.pos(comp)+(jdir)*(phi[comp]-plo[comp]);
+						Real y1 = p2.pos(comp);
+						Real vy = p.rdata(comp+1);
+						comp=2;
+						p2.pos(comp) = p.pos(comp)+(kdir)*(phi[comp]-plo[comp]);
+						Real z1 = p2.pos(comp);
+						Real vz = p.rdata(comp+1);
+
+						// There are two calls to this routine. Do the particle output writing only once 
+						// based on the bool.
+						if(is_file_write) {
+							LightConeParticle tmp;
+							tmp.x = x1; tmp.y = y1; tmp.z = z1;
+							tmp.vx = vx; tmp.vy = vy; tmp.vz = vz;
+							shell_particles.emplace_back(tmp);
+						}
+
+						for (int comp=0; comp < nc; ++comp) {
+							p2.rdata(comp+1+3)=p.pos(comp);
+							p2.rdata(comp+1+3+3) = p.pos(comp) + dt_a_cur_inv * p.rdata(comp+1);
+							p2.rdata(comp+1)=p.rdata(comp+1);
+							//              p2.pos(comp)=p.pos(comp);
+							p2.id()=p.id();
+							p2.cpu()=p.cpu();
+						}
+						return;
+					}
 				}
-
-           for (int comp=0; comp < nc; ++comp) {
-               p2.rdata(comp+1+3)=p.pos(comp);
-               p2.rdata(comp+1+3+3) = p.pos(comp) + dt_a_cur_inv * p.rdata(comp+1);
-               p2.rdata(comp+1)=p.rdata(comp+1);
-               //              p2.pos(comp)=p.pos(comp);
-               p2.id()=p.id();
-               p2.cpu()=p.cpu();
-           }
-	   return;
-	      }
-		    }
-         }
+			}
+		}
+	}
 
 }
