+++ Exec/ParticleFilterTest/Make.package
@@ -1,3 +1,4 @@
 CEXE_sources += Prob.cpp
 CEXE_sources += Prob_error.cpp
-CEXE_headers += temp.H
+CEXE_sources += LightConeParticle.cpp
+CEXE_headers += LightConeParticle.H
